# Improve perf of ReplaceErrors