Time to create 500 000 simple entities and set 3 properties each (where one property has no validation) 

Elapsed time : 5,6 seconds => 4,4 sseconds
Memory allocation 5093 MB => 3178 MB

|  | Elapsed time | RAM|
|-|-|-|
| original | 5,6s | 5093 MB |
| commit 6 - ReplaceErrors and Union => ConCat changes  | 4.4s | 3178 MB |
| commit 7 - Use HashSet instead of LINQ in OnCollectionChange  | < 4.0s | 2317 MB |